Many small changes: Reduced coupling between chunk manager and terrain generation. Split ShaderDataBuffer into a dynamic and static buffer. Changed and improved the method of rendering debug lines. Prepared for the introduction of Biomes. Minor fixed.

+++ src/voxel/Block.hpp
@@ -7,7 +7,7 @@ namespace zore {
 	static constexpr ushort FLUID_BIT   = 1 << 14;
 	static constexpr ushort OPAQUE_BIT  = 1 << 14;
 	static constexpr ushort OPAQUE_MASK = BLOCK_BIT | OPAQUE_BIT;
-	static constexpr uint   ID_MASK     = 65535u >> 2;
+	static constexpr ushort ID_MASK     = 65535u >> 2;
 
 	// Blocks
 	static constexpr ushort BLOCK_DIRT  = 0 | BLOCK_BIT | OPAQUE_BIT;
@@ -20,7 +20,7 @@ namespace zore {
 	static constexpr ushort FLUID_LAVA  = 1 | FLUID_BIT;
 
 	// Sprites and Other
-	static constexpr ushort BLOCK_AIR    = 0;
-	static constexpr ushort SPRITE_PLANT = 1;
+	static constexpr ushort BLOCK_AIR       = 0;
+	static constexpr ushort SPRITE_PLANT    = 1;
 	static constexpr ushort SPRITE_MUSHROOM = 2;
 }
